Improve support for shared libraries
The libraries are now returned as a new object type (PhotoLibrary). That
provides access to albums. This way the icloud-photos-downloader can also use
the Recently Deleted special album to sync deletions.

The list of non-special albums is currently returned empty for the
shared libraries. This also tracks with the fact that the iCloud web app
currently doesn't provide access to albums in shared libraries.